Refactoring and updates
