+++ lib/screens/common_widgets/common_widgets.dart
@@ -10,6 +10,7 @@ export 'envvar_indicator.dart';
 export 'envvar_span.dart';
 export 'envvar_popover.dart';
 export 'env_trigger_options.dart';
+export 'field_header.dart';
 export 'sidebar_filter.dart';
 export 'sidebar_header.dart';
 export 'sidebar_save_button.dart';
