+++ lib/widgets/widgets.dart
@@ -30,13 +30,13 @@ export 'editor.dart';
 export 'error_message.dart';
 export 'field_cell_obscurable.dart';
 export 'field_cell.dart';
-export 'field_header.dart';
 export 'field_json_search.dart';
 export 'field_read_only.dart';
 export 'field_url.dart';
 export 'intro_message.dart';
 export 'json_previewer.dart';
 export 'markdown.dart';
+export 'menu_header_suggestions.dart';
 export 'menu_item_card.dart';
 export 'menu_sidebar_top.dart';
 export 'overlay_widget.dart';
@@ -52,7 +52,6 @@ export 'splitview_drawer.dart';
 export 'splitview_dashboard.dart';
 export 'splitview_equal.dart';
 export 'splitview_history.dart';
-export 'suggestions_menu_box.dart';
 export 'tabbar_segmented.dart';
 export 'table_map.dart';
 export 'table_request_form.dart';
