+++ packages/apidash_design_system/lib/widgets/widgets.dart
@@ -5,5 +5,6 @@ export 'checkbox.dart';
 export 'dropdown.dart';
 export 'popup_menu.dart';
 export 'snackbar.dart';
+export 'suggestions_menu_box.dart';
 export 'textfield_outlined.dart';
 export 'textfield_raw.dart';
