+++ test/widgets/field_header_test.dart
@@ -1,8 +1,9 @@
+import 'package:apidash/screens/common_widgets/field_header.dart';
+import 'package:apidash/widgets/menu_header_suggestions.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_portal/flutter_portal.dart';
 import 'package:extended_text_field/extended_text_field.dart';
-import 'package:apidash/widgets/field_header.dart';
 import 'package:spot/spot.dart';
 
 void main() {
